requirements.txt
attrs==22.1.0
black==22.6.0
Cerberus==1.3.4
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
coverage==6.4.3
distlib==0.3.5
filelock==3.8.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
identify==2.5.3
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy==0.971
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
SQLAlchemy==1.4.40
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.11
virtualenv==20.16.3
Werkzeug==2.2.2
zipp==3.8.1
